Remove correct event in useEventSource (#119)
Small fix for the useEventSource hook: In the clean up function from the
useEffect eventSource.removeEventListener is called with the default
event "message". But if the user is using a different event name, it is
not removed. So using the same parameter as in the addEventListener call
solves that issue.